+++ src/test/java/org/opensearch/knn/index/KNNCircuitBreakerIT.java
@@ -5,11 +5,26 @@
 
 package org.opensearch.knn.index;
 
-import org.junit.Assert;
-import org.opensearch.knn.KNNRestTestCase;
+import org.apache.hc.client5.http.auth.AuthScope;
+import org.apache.hc.client5.http.auth.UsernamePasswordCredentials;
+import org.apache.hc.client5.http.impl.auth.BasicCredentialsProvider;
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManager;
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManagerBuilder;
+import org.apache.hc.client5.http.ssl.ClientTlsStrategyBuilder;
+import org.apache.hc.client5.http.ssl.NoopHostnameVerifier;
+import org.apache.hc.core5.http.HttpHost;
 import org.apache.hc.core5.http.io.entity.EntityUtils;
+import org.apache.hc.core5.http.nio.ssl.TlsStrategy;
+import org.apache.hc.core5.ssl.SSLContextBuilder;
+import org.junit.Assert;
+import org.junit.Assume;
+import org.opensearch.knn.KNNRestTestCase;
+import org.opensearch.client.Request;
 import org.opensearch.client.Response;
+import org.opensearch.client.RestClient;
 import org.opensearch.common.settings.Settings;
+import org.opensearch.common.xcontent.XContentFactory;
+import org.opensearch.core.rest.RestStatus;
 import org.opensearch.knn.index.query.KNNQueryBuilder;
 import org.opensearch.knn.plugin.stats.StatNames;
 import org.opensearch.knn.common.annotation.ExpectRemoteBuildValidation;
@@ -17,7 +32,9 @@ import org.opensearch.knn.common.annotation.ExpectRemoteBuildValidation;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 import static org.opensearch.knn.index.KNNCircuitBreaker.CB_TIME_INTERVAL;
 
@@ -28,6 +45,9 @@ public class KNNCircuitBreakerIT extends KNNRestTestCase {
     private static final Integer ALWAYS_BUILD_GRAPH = 0;
     private static final String INDEX_1 = INDEX_NAME + "1";
     private static final String INDEX_2 = INDEX_NAME + "2";
+    private static final String SECURITY_API_PREFIX = "/_plugins/_security/api";
+    private static final String SECURITY_AUTH_INFO_ENDPOINT = "/_plugins/_security/authinfo";
+    private static final String LIMITED_USER_PASSWORD = "TestUser123!";
 
     /**
      * Base setup for all circuit breaker tests.
@@ -159,6 +179,50 @@ public class KNNCircuitBreakerIT extends KNNRestTestCase {
         testNodeLevelCircuitBreaker();
     }
 
+    public void testCbTrippedWithLimitedUserWhenSecurityEnabled() throws Exception {
+        Assume.assumeTrue("Security must be enabled for this test", isSecurityEnabled());
+
+        final String testMethodSuffix = testName.getMethodName().toLowerCase(Locale.ROOT);
+        final String roleName = "knn_cb_role_" + testMethodSuffix;
+        final String userName = "knn_cb_user_" + testMethodSuffix;
+
+        createLimitedUser(roleName, userName, INDEX_1, INDEX_2);
+
+        try (RestClient limitedUserClient = createLimitedUserClient(userName)) {
+            setupIndices();
+            updateClusterSettings("knn.memory.circuit_breaker.limit", "1kb");
+
+            float[] qvector = { 1.9f, 2.4f };
+            for (int i = 0; i < 15; i++) {
+                for (String indexName : new String[] { INDEX_1, INDEX_2 }) {
+                    Request request = new Request("POST", "/" + indexName + "/_search");
+                    request.setJsonEntity(
+                        XContentFactory.jsonBuilder()
+                            .startObject()
+                            .startObject("query")
+                            .startObject("knn")
+                            .startObject(FIELD_NAME)
+                            .array("vector", qvector[0], qvector[1])
+                            .field("k", 10)
+                            .endObject()
+                            .endObject()
+                            .endObject()
+                            .endObject()
+                            .toString()
+                    );
+                    Response response = limitedUserClient.performRequest(request);
+                    assertEquals(RestStatus.OK.getStatus(), response.getStatusLine().getStatusCode());
+                }
+            }
+
+            assertBusy(() -> assertTrue("Circuit breaker should trip even for a non-admin user", isCbTripped()), 30, TimeUnit.SECONDS);
+        } finally {
+            adminClient().performRequest(new Request("DELETE", SECURITY_API_PREFIX + "/internalusers/" + userName));
+            adminClient().performRequest(new Request("DELETE", SECURITY_API_PREFIX + "/rolesmapping/" + roleName));
+            adminClient().performRequest(new Request("DELETE", SECURITY_API_PREFIX + "/roles/" + roleName));
+        }
+    }
+
     public void verifyCbUntrips() throws Exception {
 
         if (!isCbTripped()) {
@@ -174,5 +238,74 @@ public class KNNCircuitBreakerIT extends KNNRestTestCase {
             Thread.sleep(backOffInterval * 1000);
         }
         assertFalse(isCbTripped());
+    }
+
+    private boolean isSecurityEnabled() {
+        try {
+            Response response = adminClient().performRequest(new Request("GET", SECURITY_AUTH_INFO_ENDPOINT));
+            return response.getStatusLine().getStatusCode() == RestStatus.OK.getStatus();
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    private void createLimitedUser(String roleName, String userName, String... indices) throws IOException {
+        Request roleRequest = new Request("PUT", SECURITY_API_PREFIX + "/roles/" + roleName);
+        roleRequest.setJsonEntity(
+            XContentFactory.jsonBuilder()
+                .startObject()
+                .array("cluster_permissions", "cluster_composite_ops_ro")
+                .startArray("index_permissions")
+                .startObject()
+                .array("index_patterns", indices)
+                .array("allowed_actions", "read")
+                .endObject()
+                .endArray()
+                .endObject()
+                .toString()
+        );
+        assertOK(adminClient().performRequest(roleRequest));
+
+        Request userRequest = new Request("PUT", SECURITY_API_PREFIX + "/internalusers/" + userName);
+        userRequest.setJsonEntity(
+            XContentFactory.jsonBuilder()
+                .startObject()
+                .field("password", LIMITED_USER_PASSWORD)
+                .array("backend_roles", roleName)
+                .endObject()
+                .toString()
+        );
+        assertOK(adminClient().performRequest(userRequest));
+
+        Request mappingRequest = new Request("PUT", SECURITY_API_PREFIX + "/rolesmapping/" + roleName);
+        mappingRequest.setJsonEntity(
+            XContentFactory.jsonBuilder().startObject().array("users", userName).array("backend_roles", roleName).endObject().toString()
+        );
+        assertOK(adminClient().performRequest(mappingRequest));
+    }
+
+    private RestClient createLimitedUserClient(String userName) {
+        BasicCredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+        credentialsProvider.setCredentials(
+            new AuthScope(null, -1),
+            new UsernamePasswordCredentials(userName, LIMITED_USER_PASSWORD.toCharArray())
+        );
+        try {
+            TlsStrategy tlsStrategy = ClientTlsStrategyBuilder.create()
+                .setHostnameVerifier(NoopHostnameVerifier.INSTANCE)
+                .setSslContext(SSLContextBuilder.create().loadTrustMaterial(null, (chains, authType) -> true).build())
+                .build();
+            PoolingAsyncClientConnectionManager connectionManager = PoolingAsyncClientConnectionManagerBuilder.create()
+                .setTlsStrategy(tlsStrategy)
+                .build();
+            return RestClient.builder(getClusterHosts().toArray(new HttpHost[0]))
+                .setHttpClientConfigCallback(
+                    httpClientBuilder -> httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider)
+                        .setConnectionManager(connectionManager)
+                )
+                .build();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 }